Engine - Se cambio el nombre de Ground a CollisionMap - Se cambio los parametors de entrada para crear entidades, ahora se requiere posicion, dimension y un mapa asociativo de propiedades en vez de parametros variables. - Los loaders fueron movidos al paquete loaders - Se añadio el recurso Level con su respectivo loader.

+++ Apollo/src/com/apollo/behaviors/GunBehavior.java
@@ -106,7 +106,7 @@ public final class GunBehavior extends BaseBehavior implements GunMessageHandler
 			float r = transform.rotation + rot*MathUtils.degreesToRadians;			
 			final float sin = MathUtils.sin(r);
 			final float cos = MathUtils.cos(r);
-			Entity gunEnt = world.createEntity(entity,transform.x + dist*cos,transform.y +dist*sin, vel*cos, vel*sin);
+			Entity gunEnt = world.createEntity(entity,(int)(transform.x + dist*cos),(int)(transform.y +dist*sin),(int)(vel*cos), (int)(vel*sin));
 			world.addEntity(gunEnt);
 
 			ammo -=1;
